dsi: count the quantum against write data; align it to the TCP MSS
Problems: dsi_stream_receive bounded the whole DSI frame by the server
quantum, so a spec-legal FPWrite carrying exactly quantum data bytes
(dsi_len = quantum + doff) tore down the session. To stay under that
bound the tools subtracted the frame overhead from every transfer, so
clients never used the full negotiated quantum in either direction --
a throughput loss on both writes and reads from incorrect header
accounting, not a real protocol limit. dsi_writeinit copied the
buffered payload prefix into the caller's buffer (a second userspace
copy of up to the read-ahead size per write) and silently truncated
when the prefix exceeded the caller's buffer. The test client sent
each write frame as two syscalls, leaving a runt TCP segment before
every payload; transfers also ended in a runt segment whenever the
quantum was not aligned to the path MSS.

Model: per the AFP spec (DSIOpenSession, Table 1: the request quantum
is the write data size 'not including the length of the DSI header and
the AFP command'), the server quantum counts FPWrite/FPWriteExt data
only; the 16-byte DSI header and the request block ride on top. The
session quantum is resolved once per session after accept(): the
smallest 4 KiB multiple at or above the configured value whose
FPWriteExt frame (data + 36 bytes) ends in a TCP segment at least 7/8
full, growing by at most 256 KiB, capped at 256 MiB; best fill wins
when the threshold is unreachable. Buffers, inbound checks, and the
advertised value all use the grown quantum, so advertised == capacity.

Fixes: writes are bounded by dsi_len - doff with doff validated to
{12, 20}; non-write commands keep the strict bound. This restores
full-quantum transfers -- a client may again send (and request) a
complete server_quantum of data per frame, recovering the throughput
the overhead-subtraction workarounds gave up. dsi_writeinit hands the
buffered payload back by reference (NULL drains it) -- one copy, no
truncation, and afp_addicon's error path drains and returns instead of
re-arming the transfer. The client sends header, request block, and
payload in a single writev; speedtest selects FPWriteExt or legacy
FPWrite automatically from the negotiated AFP version (FPWriteExt for
AFP >= 3.0), and both tools size transfers from the full advertised
quantum with no overhead subtraction and no zero-quantum division. dsireadbuf defaults to 32, capped so dsireadbuf
x quantum stays under 1 GiB per session; server quantum values above
the maximum clamp instead of falling back to the default.
dsi_writeinit's signature change bumps the soversion to 21.

Security: the DSI frame-parsing hardening from CVE-2026-62319 and
CVE-2026-62320 (commit cc0a9d16) stays fully in place and is
strengthened -- the write data offset is now validated against an
exact {12, 20} whitelist, so malformed offsets 13-19 are rejected
instead of parsed, and dsi_tcp_open's first-frame whole-frame bound is
unchanged. Those fixes correctly reject a frame whose total exceeds
the quantum; the tools had compensated with frame-overhead
subtraction (PR #3186, PR #3187), which shrank the usable quantum.
This supersedes those workarounds by fixing the server semantics they
compensated for -- restoring full-quantum throughput -- while
preserving their invariant that the client never emits a frame the
hardened parser rejects.

Additionally closes an out-of-bounds read in the FPAddIcon write path:
the old dsi_writeinit truncated the buffered prefix into the caller's
buffer via memmove(buf, start, MIN(buflen, bytes)) but returned the
untruncated byte count, and afp_addicon then wrote that count from a
64 KiB (DSI_DATASIZ) buffer -- reading past the buffer and emitting
stale heap bytes into the icon file whenever a buffered write prefix
exceeded 64 KiB against the 1 MiB default quantum. The pointer handoff
returns the exact buffered range, removing the over-read.

Tests: afpd unit tests pin full-quantum accept (both doffs),
over-quantum reject, doff validation, the strict non-write bound, the
no-duplicate-copy handoff, the alignment search (growth, page
alignment, caps, overflow at the top of the range), and the read-ahead
clamp. RED runs for the two server bugs are preserved in the PR's CI
history. The speedtest dashboard gains sweep-average columns so
small-file regressions are visible alongside peaks.

+++ test/testsuite/speedtest_local_vfs.c
@@ -312,21 +312,22 @@ uint16_t local_openfork(CONN *conn _U_, uint16_t vol, int type,
 /*!
  * @brief Write data via lseek() + write() syscalls
  */
-unsigned int local_writeheader(DSI *dsi _U_, uint16_t fork, int offset,
-                               int size,
+unsigned int local_writeheader(DSI *dsi _U_, uint16_t fork, off_t offset,
+                               size_t size,
                                char *data, char whence)
 {
     if (!Local_VFS_Quiet) {
         fprintf(stdout, "---------------------\n");
-        fprintf(stdout, "send write header fork %d  offset %d size %d from 0x%x\n\n",
-                fork, offset, size, (unsigned)whence);
+        fprintf(stdout,
+                "send write header fork %d  offset %lld size %zu from 0x%x\n\n",
+                fork, (long long)offset, size, (unsigned)whence);
     }
 
     if (lseek(fork, offset, SEEK_SET) == (off_t) -1) {
         return ntohl(AFPERR_EOF);
     }
 
-    if (write(fork, data, size) != size) {
+    if (write(fork, data, size) != (ssize_t)size) {
         return ntohl(AFPERR_EOF);
     }
 
@@ -336,14 +337,15 @@ unsigned int local_writeheader(DSI *dsi _U_, uint16_t fork, int offset,
 /*!
  * @brief No-op write footer (AFP compatibility)
  */
-unsigned int local_writefooter(DSI *dsi _U_, uint16_t fork, int offset,
-                               int size,
+unsigned int local_writefooter(DSI *dsi _U_, uint16_t fork, off_t offset,
+                               size_t size,
                                char *data _U_, char whence)
 {
     if (!Local_VFS_Quiet) {
         fprintf(stdout, "---------------------\n");
-        fprintf(stdout, "get write footer fork %d  offset %d size %d from 0x%x\n\n",
-                fork, offset, size, (unsigned)whence);
+        fprintf(stdout,
+                "get write footer fork %d  offset %lld size %zu from 0x%x\n\n",
+                fork, (long long)offset, size, (unsigned)whence);
     }
 
     return ntohl(AFP_OK);
@@ -452,13 +454,14 @@ unsigned int local_read(CONN *conn _U_, uint16_t fork, long long offset,
 /*!
  * @brief No-op read header (AFP compatibility)
 */
-unsigned int local_readheader(DSI *dsi _U_, uint16_t fork, int offset, int size,
+unsigned int local_readheader(DSI *dsi _U_, uint16_t fork, off_t offset,
+                              size_t size,
                               char *data _U_)
 {
     if (!Local_VFS_Quiet) {
         fprintf(stdout, "---------------------\n");
-        fprintf(stdout, "send read header fork %d  offset %d size %d\n\n", fork, offset,
-                size);
+        fprintf(stdout, "send read header fork %d  offset %lld size %zu\n\n", fork,
+                (long long)offset, size);
     }
 
     return ntohl(AFP_OK);
@@ -467,20 +470,21 @@ unsigned int local_readheader(DSI *dsi _U_, uint16_t fork, int offset, int size,
 /*!
  * @brief Read data footer via lseek() + read()
 */
-unsigned int local_readfooter(DSI *dsi _U_, uint16_t fork, int offset, int size,
+unsigned int local_readfooter(DSI *dsi _U_, uint16_t fork, off_t offset,
+                              size_t size,
                               char *data)
 {
     if (!Local_VFS_Quiet) {
         fprintf(stdout, "---------------------\n");
-        fprintf(stdout, "get read reply fork %d  offset %d size %d\n\n", fork, offset,
-                size);
+        fprintf(stdout, "get read reply fork %d  offset %lld size %zu\n\n", fork,
+                (long long)offset, size);
     }
 
     if (lseek(fork, offset, SEEK_SET) == (off_t) -1) {
         return ntohl(AFPERR_EOF);
     }
 
-    if (read(fork, data, size) != size) {
+    if (read(fork, data, size) != (ssize_t)size) {
         return ntohl(AFPERR_EOF);
     }
